refactor: version 2
